requirements.txt
# Kotak Neo 5-Minute Volatility Harvester
# Production dependencies — pinned for reproducibility

# Kotak Neo v2 SDK (authentication + WebSocket)
# Install from GitHub: pip install "git+https://github.com/Kotak-Neo/Kotak-neo-api-v2.git@v2.0.1#egg=neo_api_client"
neo_api_client @ git+https://github.com/Kotak-Neo/Kotak-neo-api-v2.git@v2.0.1

# Google Sheets API
gspread>=5.12.0
oauth2client>=4.1.3
google-auth>=2.23.0

# TOTP generation for 2FA
pyotp>=2.9.0

# Environment configuration
python-dotenv>=1.0.0

# Scheduling
APScheduler>=3.10.4

# Timezone support
pytz>=2024.1

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
numpy>=1.24.0
